First Approach - Additive Hash
Uses a loop to convert characters to ASCII and add to the running total to calculate the hash. The results for this approach showed a large number of collisions with a fast build time.  A lot of buckets are unused showing a largely empty table. Overall the approach is extremely simple but not very effective