Approach 5 - XOR Hashing
My intent with this approach was to try and make the build time as fast as I could. XOR hashing is meant to be extremely fast which is was for both tables. I wasn't however expecting it to roughly match the amount of collisions and wasted buckets as the last 2 approaches